Update README.md
This C program generates a shuffled array based on user input. It prompts the user to enter the number of elements in the array and then takes input of the array elements. After that, it shuffles the elements of the array using the Fisher-Yates Shuffle algorithm and prints the shuffled array to the console.